LayoutTests/imported/w3c/web-platform-tests/css/css-mixins/function-in-media-expected.txt



FAIL Custom Functions: @function inside @media assert_equals: --actual before resize expected "50" but got "--f()"
